// Frontend/src/app/core/models/user.model.ts
export interface User {
  id: string;
  name: string;
  email: string;
  phone?: string;
  token?: string;
}

export interface LoginPayload {
  email: string;
  password: string;
}

export interface RegisterPayload {
  name: string;
  email: string;
  password: string;
  phone?: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}